requirements.txt
streamlit
huggingface_hub
python-dotenv
requests
yfinance
matplotlib
pandas
sentence-transformers
faiss-cpu
